macOS: combine license + Cloud credits into one account panel
Merge the separate License and HyperWhisper Cloud Credits settings
sections into a single always-visible "HyperWhisper Cloud" panel, since
the license key is now the wallet that holds Cloud credits.

- New CloudAccountSettingsSection replaces CreditsSettingsSection and
  LicenseSettingsSection.
- Licensed view keeps the existing credits card style and adds the
  license key as a masked (password-style) row with reveal + copy.
- Single "Purchase More" button opens the web /credits page (no in-app
  tier picker); unlicensed users get a key-entry block + "Get Credits".
- SettingsView collapses the two sidebar entries into one (tag "license"
  preserved for existing deep links); drops the deactivate redirect.
- Adds settings.cloud.* strings to Base.lproj.
