+++ app/macos/hyperwhisper/Localizations/Base.lproj/Localizable.strings
@@ -532,6 +532,19 @@
 "settings.credits.error.network" = "Network error: %@";
 "settings.credits.error.invalidResponse" = "Invalid response from server. Please try again.";
 "settings.credits.error.server" = "Server error: %@";
+
+// MARK: - Settings · Cloud (combined license + credits)
+"settings.section.cloud" = "HyperWhisper Cloud";
+"settings.cloud.licenseKey" = "License Key";
+"settings.cloud.purchaseMore" = "Purchase More";
+"settings.cloud.key.reveal" = "Reveal license key";
+"settings.cloud.key.hide" = "Hide license key";
+"settings.cloud.key.copy" = "Copy license key";
+"settings.cloud.activate.title" = "Activate HyperWhisper Cloud";
+"settings.cloud.activate.subtitle" = "Your license key is your wallet — it holds your Cloud credits. Enter a key to activate, or buy credits to get one.";
+"settings.cloud.getCredits.title" = "Don't have a key yet?";
+"settings.cloud.getCredits.subtitle" = "Buy credits and we'll email you a license key.";
+"settings.cloud.getCredits.button" = "Get Credits";
 "settings.models.alert.limit.title" = "Model download limit reached";
 "settings.models.alert.limit.message" = "Trials include %d model downloads. Upgrade your license for unlimited downloads.";
 "settings.models.alert.cannotDelete.title" = "Cannot Delete Model";
