+++ app/macos/hyperwhisper/Views/SettingsView.swift
@@ -42,23 +42,13 @@ struct SettingsView: View {
             Label("settings.section.sound".localized, systemImage: "speaker.wave.2")
                 .tag("sound")
 
-            Label("settings.section.license".localized, systemImage: "checkmark.seal")
+            // COMBINED LICENSE + CLOUD CREDITS
+            // The license key is the wallet, so license activation and the credit
+            // balance now live in one always-visible section (CloudAccountSettingsSection).
+            // Tagged "license" so existing deep links (e.g. MainAppView "Enter License Key")
+            // still resolve here.
+            Label("settings.section.cloud".localized, systemImage: "cloud")
                 .tag("license")
-
-            // CONDITIONAL RENDERING: Credits section only visible for licensed users
-            // Trial users (using device_id) don't have access to HyperWhisper Cloud credits
-            // Only licensed users (using license_key) get the $5 credit allocation
-            //
-            // Why this matters:
-            // 1. Trial users use device_id for identification → no persistent credit pool
-            // 2. Licensed users use license_key → get $5 credit allocation
-            // 3. Showing credits tab to trial users would be confusing (nothing to display)
-            // 4. When user activates license, this section appears automatically
-            // 5. When user deactivates license, this section disappears and redirects to General
-            if licenseManager.licenseStatus == .active {
-                Label("settings.section.credits".localized, systemImage: "creditcard.fill")
-                    .tag("credits")
-            }
 
             Label("settings.section.storage".localized, systemImage: "folder")
                 .tag("storage")
@@ -79,24 +69,6 @@ struct SettingsView: View {
                 .tag("apiserver")
         }
         .listStyle(.sidebar)
-        .onChange(of: licenseManager.licenseStatus) { oldStatus, newStatus in
-            // AUTO-NAVIGATION ON LICENSE CHANGE
-            // If the user is currently viewing Credits and they deactivate their license,
-            // we need to redirect them to another section since Credits will disappear
-            //
-            // Flow:
-            // 1. User is viewing Credits section (selectedSettingsSection == "credits")
-            // 2. User clicks "Deactivate Device" in License section
-            // 3. License status changes from .active → .trial
-            // 4. Credits section disappears from sidebar (conditional rendering)
-            // 5. This onChange detects the status change
-            // 6. If currently viewing credits, redirect to General settings
-            // 7. Prevents "blank screen" UX issue
-            if oldStatus == .active && newStatus != .active && appState.selectedSettingsSection == "credits" {
-                appState.selectedSettingsSection = "general"
-                AppLogger.ui.info("License deactivated while viewing Credits · redirecting to General settings")
-            }
-        }
     }
 
     @ViewBuilder
@@ -104,10 +76,8 @@ struct SettingsView: View {
         switch identifier {
         case "sound":
             SoundSettingsSection()
-        case "license":
-            LicenseSettingsSection()
-        case "credits":
-            CreditsSettingsSection()
+        case "license", "credits":
+            CloudAccountSettingsSection()
         case "storage":
             StorageSettingsSection()
         case "shortcuts":
